Style the famous oils section
