+++ client/src/components/organisms/FamousOils/FamousOils.scss
@@ -1,0 +1,7 @@
+.famous-oils {
+    margin-bottom: 5rem;
+}
+
+.famous-oils__title {
+    font-size: var(--text-md);
+}
