+++ client/src/components/pages/Homepage/Homepage.jsx
@@ -1,40 +1,18 @@
 import { useEffect, useState } from "react";
-import { IconArrowLeft } from '@tabler/icons-react';
 import Search from '../../molecules/Search/Search';
 import Categories from "../../molecules/Categories/Categories";
 import MostWanted from "../../molecules/MostWanted/MostWanted";
+import FamousOils from "../../organisms/FamousOils/FamousOils";
 
 
 function Homepage() {
-    const [oils, setOils] = useState(null);
-
-    useEffect(() => {
-        const fetchOils = async () => {
-            const response = await fetch('/api/oils');
-            const json = await response.json();
-
-            if(response.ok) {
-                setOils(json);
-            }
-        }
-
-        fetchOils();
-    }, []);
-
+    
     return (
         <div className='Homepage'>
             <Search />
             <Categories />
             <MostWanted />
-            {/* <div>
-                {oils && oils.map(oil => (
-                    <div key={oil.id}>
-                        <h2>{oil.name}</h2>
-                        <p>{oil.description}</p>
-                    </div>
-                ))}    
-            </div>
-                <IconArrowLeft /> */}
+            <FamousOils />
         </div>
     );
 }
